+++ internal/domain/repository/room_repository.go
@@ -11,7 +11,7 @@ type RoomRepository interface {
 	Create(ctx context.Context, room *entity.CreateRoom) error
 
 	// GetByID retrieves a room by ID
-	GetByID(ctx context.Context, id uint) (*entity.Room, error)
+	GetByID(ctx context.Context, id uint64) (*entity.Room, error)
 
 	// List retrieves rooms based on filter
 	List(ctx context.Context, filter *entity.RoomFilter) ([]entity.Room, int64, error)
@@ -20,7 +20,7 @@ type RoomRepository interface {
 	Update(ctx context.Context, room *entity.Room) error
 
 	// Delete deletes a room by ID
-	Delete(ctx context.Context, id uint) error
+	Delete(ctx context.Context, id uint64) error
 }
 
 // RoomCategoryRepository defines the interface for room category data operations
