+++ internal/domain/usecase/user_usecase.go
@@ -9,16 +9,16 @@ import (
 // UserUseCase defines the interface for user business logic
 type UserUseCase interface {
 	// GetUserByID retrieves a user by ID
-	GetUserByID(ctx context.Context, id uint) response.StatusResponse
+	GetUserByID(ctx context.Context, id uint64) response.StatusResponse
 
 	// GetListUsers retrieves users based on filter
 	GetListUsers(ctx context.Context, filter *entity.UserFilter) response.StatusResponse
 
 	// UpdateUser updates a user
-	UpdateUser(ctx context.Context, id uint, user *entity.User) response.StatusResponse
+	UpdateUser(ctx context.Context, id uint64, user *entity.User) response.StatusResponse
 
 	// DeleteUser deletes a user
-	DeleteUser(ctx context.Context, id uint) response.StatusResponse
+	DeleteUser(ctx context.Context, id uint64) response.StatusResponse
 
 	// AddUserToRoom adds a user to a room
 	AddUserToRoom(ctx context.Context, payload entity.AddUserToRoom) response.StatusResponse
