+++ internal/usecase/user_usecase.go
@@ -27,7 +27,7 @@ func NewUserUseCase(repos repository.Repositories, logger logger.Logger) usecase
 }
 
 // GetUserByID retrieves a user by ID
-func (uc *userUseCase) GetUserByID(ctx context.Context, id uint) response.StatusResponse {
+func (uc *userUseCase) GetUserByID(ctx context.Context, id uint64) response.StatusResponse {
 	user := &entity.User{
 		Base: entity.Base{ID: id},
 	}
@@ -51,7 +51,7 @@ func (uc *userUseCase) GetListUsers(ctx context.Context, filter *entity.UserFilt
 }
 
 // UpdateUser updates a user
-func (uc *userUseCase) UpdateUser(ctx context.Context, id uint, userData *entity.User) response.StatusResponse {
+func (uc *userUseCase) UpdateUser(ctx context.Context, id uint64, userData *entity.User) response.StatusResponse {
 	user := &entity.User{
 		Base: entity.Base{ID: id},
 	}
@@ -89,7 +89,7 @@ func (uc *userUseCase) UpdateUser(ctx context.Context, id uint, userData *entity
 }
 
 // DeleteUser deletes a user
-func (uc *userUseCase) DeleteUser(ctx context.Context, id uint) response.StatusResponse {
+func (uc *userUseCase) DeleteUser(ctx context.Context, id uint64) response.StatusResponse {
 	user := &entity.User{
 		Base: entity.Base{ID: id},
 	}
